requirements.txt
boto3==1.28.17
diffusers==0.16.1
matplotlib==3.7.1
pillow==9.4.0
pyunsplash==1.0.0rc2
pyyaml==6.0
requests==2.31.0
safetensors==0.3.3
sagemaker==2.179.0
selenium==4.11.2
torch==1.13.1
transformers==4.32.0
tqdm==4.65.0
webdriver-manager==4.0.0